# Inline based optimizations