--- a/src/Control/Monad/Logic/Sequence.hs
+++ b/src/Control/Monad/Logic/Sequence.hs
@@ -101,6 +101,7 @@ instance TASequence s => T.Traversable (MSeq s) where
 
 fromView :: m (View m a) -> SeqT m a
 fromView = SeqT . singleton
+{-# INLINE fromView #-}
 
 toView :: Monad m => SeqT m a -> m (View m a)
 toView (SeqT s) = case viewl s of
@@ -108,18 +109,25 @@ toView (SeqT s) = case viewl s of
   h S.:< t -> h >>= \x -> case x of
     Empty -> toView (SeqT t)
     hi :< SeqT ti -> return (hi :< SeqT (ti S.>< t))
+{-# INLINE toView #-}
 
 single :: Monad m => a -> m (View m a)
 single a = return (a :< mzero)
+{-# INLINE single #-}
 
 instance Monad m => Functor (SeqT m) where
+  {-# INLINEABLE fmap #-}
   fmap f xs = xs >>= return . f
 
 instance Monad m => Applicative (SeqT m) where
+  {-# INLINE pure #-}
+  {-# INLINEABLE (<*>) #-}
   pure = fromView . single
   (<*>) = liftM2 id
 
 instance Monad m => Alternative (SeqT m) where
+  {-# INLINE empty #-}
+  {-# INLINEABLE (<|>) #-}
   empty = SeqT (MSeq tempty)
   (toView -> m) <|> n = fromView (m >>= \x -> case x of
       Empty -> toView n
@@ -127,6 +135,8 @@ instance Monad m => Alternative (SeqT m) where
     where cat (SeqT l) (SeqT r) = SeqT (l S.>< r)
 
 instance Monad m => Monad (SeqT m) where
+  {-# INLINE return #-}
+  {-# INLINEABLE (>>=) #-}
   return = fromView . single
   (toView -> m) >>= f = fromView (m >>= \x -> case x of
     Empty -> return Empty
@@ -139,21 +149,29 @@ instance Monad m => Fail.MonadFail (SeqT m) where
   fail _ = SeqT S.empty
 
 instance Monad m => MonadPlus (SeqT m) where
+  {-# INLINE mzero #-}
+  {-# INLINE mplus #-}
   mzero = Control.Applicative.empty
   mplus = (<|>)
 
 #if MIN_VERSION_base(4,9,0)
 instance Monad m => Semigroup (SeqT m a) where
+  {-# INLINE (<>) #-}
+  {-# INLINE sconcat #-}
   (<>) = mplus
   sconcat = foldr1 mplus
 #endif
 
 instance Monad m => Monoid (SeqT m a) where
+  {-# INLINE mempty #-}
+  {-# INLINE mappend #-}
+  {-# INLINE mconcat #-}
   mempty = SeqT (MSeq tempty)
   mappend = (<|>)
   mconcat = F.asum
 
 instance MonadTrans SeqT where
+  {-# INLINE lift #-}
   lift m = fromView (m >>= single)
 
 instance Monad m => MonadLogic (SeqT m) where
@@ -166,8 +184,9 @@ instance Monad m => MonadLogic (SeqT m) where
 
 observeAllT :: Monad m => SeqT m a -> m [a]
 observeAllT (toView -> m) = m >>= go where
-  go (a :< t) = liftM (a:) (observeAllT t)
+  go (a :< t) = liftM (a:) (toView t >>= go)
   go _ = return []
+{-# INLINEABLE observeAllT #-}
 
 #if !MIN_VERSION_base(4,13,0)
 observeT :: Monad m => SeqT m a -> m a
@@ -177,23 +196,29 @@ observeT :: MonadFail m => SeqT m a -> m a
 observeT (toView -> m) = m >>= go where
   go (a :< _) = return a
   go Empty = fail "No results."
+{-# INLINE observeT #-}
 
 observe :: Seq a -> a
 observe (toView -> m) = case runIdentity m of
   a :< _ -> a
   Empty -> error "No results."
+{-# INLINE observe #-}
 
 observeMaybeT :: Monad m => SeqT m (Maybe a) -> m (Maybe a)
 observeMaybeT (toView -> m) = m >>= go where
   go (Just a :< _) = return (Just a)
   go (Nothing :< rest) = toView rest >>= go
   go Empty = return Nothing
+{-# INLINEABLE observeMaybeT #-}
 
 observeMaybe :: Seq (Maybe a) -> Maybe a
 observeMaybe = runIdentity . observeMaybeT
+{-# INLINE observeMaybe #-}
 
 observeAll :: Seq a -> [a]
 observeAll = runIdentity . observeAllT
+{-# INLINE observeAll #-}
 
 instance MonadIO m => MonadIO (SeqT m) where
+  {-# INLINE liftIO #-}
   liftIO = lift . liftIO