I also switched observeAllT to use a worker/wrapper
pattern.

I just sort of went nuts marking things as `INLINE` but one of the biggest payoffs was actually `>>=`.

```
SeqT IO with msplit:
Batch generated in: 0.546473 seconds
Batch generated in: 0.521952 seconds
Batch generated in: 0.506858 seconds
Batch generated in: 0.536403 seconds
Batch generated in: 0.524648 seconds
Total time: 2.636488 seconds

SeqT Identity with msplit:
Batch generated in: 0.60177 seconds
Batch generated in: 0.5683549999999999 seconds
Batch generated in: 0.6440279999999999 seconds
Batch generated in: 0.66166 seconds
Batch generated in: 0.5953309999999999 seconds
Total time: 3.071288 seconds

SeqT Identity without msplit:
Batch generated in: 0.6014269999999999 seconds
Batch generated in: 0.641728 seconds
Batch generated in: 0.623448 seconds
Batch generated in: 0.550899 seconds
Batch generated in: 0.635666 seconds
Total time: 3.053315 seconds

LogicT Identity without msplit:
Batch generated in: 0.115443 seconds
Batch generated in: 0.141348 seconds
Batch generated in: 0.130463 seconds
Batch generated in: 0.117857 seconds
Batch generated in: 0.163205 seconds
Total time: 0.668446 seconds
```