@dagit 2021-07-22
> This is going to make my PR pretty hard to rebase.

Sorry, although I think it should be easy to merge because aside from rewriting `observeAllT` all the other changes are just adding `INLINE` or `INLINEABLE` pragma. You could just prefer your version in a merge and then add them back. Something like this: https://stackoverflow.com/questions/17704119/how-to-prefer-files-from-one-branch-during-a-merge

> I don't understand what your benchmark numbers are referring to.

Yeah, I need to publish the benchmark. It's just a simple iterative deepening depth first search that generates paths and then I convert those `streamly` streams in batches of half a million. Then output them all. The streamly stuff isn't really necessary, it was just part of a real workload I had. Anyway, when I uploaded 0.1.0.0 there was a factor 10x between the last output (`LogicT`) and any of the `SeqT` outputs. Now the gap is about 4.3x.