@dagit 2021-07-22
I should add, the other implication here is that anything you can do in your rewrite to optimize `>>=` will have a pretty solid pay off.