@treeowl 2021-07-22
My big PR just reorganizes the modules (a bunch) and adds pattern synonyms. I'll just have to figure out the rebase.